refactor: move utility functions to appropriate files
