+++ utils/underscore-query.ts
@@ -1,0 +1,6 @@
+import _ from 'lodash';
+import lodashModifier from 'underscore-query';
+
+lodashModifier(_);
+
+export default _.query;
